Add CLI to checkpoint tipsets
